Check_puppet_agent: Fix bug in perfdata and change exitcode for Disabled
I forgot to change the exit code to 1 (WARNING) in case puppet is
disabled. This was suggested by Keith Gillow (see 2321ea1affc5504cd03f21109862b5c30c5c7697).

PR (#15) contains a bug: compset is a bash internal command and this
script is written for sh (for compatibility reasons). This is fixed in
this commit by replacing compset with set and grep a bit more precise.

I also replaced the "| " before the $PERF_DATA in the resultfunction
because that looked a bit awkward if -P wasn't set.

Remove trailing ", " from $PERF_DATA